# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: switchboard
  version: "1"
  summary: Coordination for parallel Claude sessions.
  description: |
    Five sessions build in parallel, a sixth hands out the work, and a PO
    decides. switchboard carries the three things GitHub cannot say: the living
    state of a session, an ask addressed to a human and followed through to its
    answer, and a place where the PO answers without a stream of text running
    over them.

    Issues, labels and pull requests stay on GitHub. Nothing here duplicates them.

    **No authentication.** The service listens on the loopback interface, on the
    machine, for the sessions on it.

    **Two calls block.** `GET /v1/events/{id}/reply` and `GET /v1/wake` take a
    `wait` in seconds, capped at ten minutes, and answer `204` when the wait runs
    out with nothing to say. The caller loops. Nothing polls the database: every
    write wakes the waits in flight.
  license:
    name: MIT
...
servers:
  - url: http://127.0.0.1:8787
    description: The installed service.
  - url: http://127.0.0.1:8799
    description: A development service (`just dev`).
...
paths:
  /v1/sessions/{name}:
    put:
      tags: [sessions]
      operationId: saveSession
      summary: Declare what a session is doing
      description: |
        Creates the session on first call. `since_at` only moves when the status
        or the issue changes, so a repeated heartbeat does not reset it: the PO's
        table must be able to say "40 min", not "just now" five times a minute.
      parameters:
        - name: name
          in: path
          required: true
          description: The session's name, e.g. `acme-dev3`. Given by the caller, never guessed.
          schema: { type: string, examples: ["acme-dev3"] }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SessionRequest" }
      responses:
        "200":
          description: The session as it now stands.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/events:
    post:
      tags: [events]
      operationId: addEvent
      summary: Publish a typed event
      description: |
        The one call a dev session makes to say something without interrupting
        anybody.

        `audience` is worked out when it is left out: a `validation` goes to the
        PO, everything else to the manager, who dispatches.

        `body` is sanitized on the way in against a short whitelist — links,
        bold, italic, code, lists, paragraphs. Anything else is kept as text
        rather than run. What is stored is already clean.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventRequest" }
      responses:
        "201":
          description: The event, with its id and its sanitized body.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400":
          description: |
            Missing `author`, `kind` or `title`; an unknown `kind`, `role` or
            `audience`; a title over one line or 120 characters; or a
            `validation` with no `link` — which is a validation nobody can give.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/events/{id}:
    get:
      tags: [events]
      operationId: getEvent
      summary: One event and its answer
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: The event.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/events/{id}/reply:
    get:
      tags: [events]
      operationId: awaitReply
      summary: Wait for the answer to an event
      description: |
        Blocks until the event has an answer, so a dev can go back to work on its
        own. Answers `204` when the wait runs out; the caller loops.

        After reading an answer, call `POST /v1/events/{id}/seen`. That is what
        tells an undo it is too late to be invisible.
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/Wait"
      responses:
        "200":
          description: The event, answer included.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "204":
          description: The wait ran out with no answer yet. Ask again.
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/events/{id}/seen:
    post:
      tags: [events]
      operationId: markSeen
      summary: Say the asker has read the answer
      description: |
        An answer that has been read cannot be taken back silently: somebody is
        already working on it. This is what makes that knowable.
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "204":
          description: Recorded.
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/events/{id}/replies:
    post:
      tags: [answers]
      operationId: addReply
      summary: Answer an event
      description: |
        An event takes **one** answer. Answering twice is a mistake, not an edit.
        To correct one, take it back with `undo` inside the window.
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReplyRequest" }
      responses:
        "201":
          description: The answer as recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reply" }
        "400":
          description: No `author`, or neither `text` nor `option`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The event already has an answer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/events/{id}/undo:
    post:
      tags: [answers]
      operationId: undoReply
      summary: Take back an answer just given
      description: |
        Ten seconds by default (`--undo-window`). Short on purpose: it catches
        the wrong button, not a change of mind. Only the author of an answer can
        take it back.

        The event becomes actionable **as of now**, not as of its creation — the
        batching must treat it as something new, or nobody would be told it is
        waiting again.

        If the asker had already read the answer, the withdrawal cannot be
        silent. It then raises an ask to the manager naming the dev to reach, and
        `catch_up_id` points at it. The service never talks to a session itself.
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                author:
                  type: string
                  description: Who wrote the answer. Defaults to `po`.
                  default: po
      responses:
        "200":
          description: The event, open again.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UndoResponse" }
        "404":
          description: No such event, or it has no answer to take back.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Past the window, or the answer was written by somebody else.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/events/{id}/reroute:
    post:
      tags: [manager]
      operationId: rerouteEvent
      summary: Change who an open event is addressed to
      description: The manager handing a dev's question over to the PO.
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [audience]
              properties:
                audience: { $ref: "#/components/schemas/Audience" }
      responses:
        "200":
          description: The event, now addressed elsewhere.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: No such event, or it is no longer open.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/wake:
    get:
      tags: [manager]
      operationId: wake
      summary: The grouped signal
      description: |
        The point of the whole tool. An idle session only picks work back up when
        something prompts it, and one signal per event would rebuild the defect
        this service exists to remove. So this call blocks until a batch is ripe
        and then answers **a count, never the content**.

        A signal is owed when, all at once: at least one item has not been
        announced yet; the oldest of those is past the batching delay (3 min, or
        30 s when a dev is blocked); and the last signal is further back than the
        floor (5 min).

        What counts as waiting: open events addressed to that party that are not
        `info`, plus unread answers to questions they asked themselves. `info`
        never counts — it exists to be published without anybody having to read
        it now.
      parameters:
        - name: audience
          in: query
          required: false
          description: Whose batch. Defaults to `manager`.
          schema: { $ref: "#/components/schemas/Audience" }
        - $ref: "#/components/parameters/Wait"
      responses:
        "200":
          description: A batch is ripe.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WakeResponse" }
        "204":
          description: Nothing is ripe. Ask again.
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  parameters:
    EventId:
      name: id
      in: path
      required: true
      description: The event's id.
      schema: { type: integer, format: int64, examples: [12] }
    Wait:
      name: wait
      in: query
      required: false
      description: |
        How long to block, in seconds. Capped at 600. Zero or absent returns
        straight away.
      schema: { type: integer, minimum: 0, maximum: 600, default: 0 }
  responses:
    BadRequest:
      description: The request does not hold together; the message says how.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such event.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: What went wrong, in terms the caller can act on.
      required: [error]
...
    Audience:
      type: string
      description: Who an event is addressed to.
      enum: [manager, po]
...
    SessionRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/SessionStatus" }
        issue: { $ref: "#/components/schemas/Issue" }
        detail:
          type: string
          description: "One short line: what is happening. Shown in the PO's table, clipped to one line."
          examples: ["action bar, label of the primary button"]
    Session:
      type: object
      properties:
        name: { type: string }
        status: { $ref: "#/components/schemas/SessionStatus" }
        issue: { $ref: "#/components/schemas/Issue" }
        detail: { type: string }
        since_at:
          type: string
          format: date-time
          description: When the session entered this status, on this issue.
        updated_at:
          type: string
          format: date-time
          description: Last sign of life.
      required: [name, status, since_at, updated_at]
...